Add files via upload
These are all the files that I have added changes to based on the initial branch that trang had uploaded not too long ago. 

I have added the css to the corresponding html files and linked them accordingly. There is some new html added. The urls are added as well. In views.py there are changes to the def shopnow, def account, and def create_order. There are also some models for a form in models.py but I am not finished with it as of yet. It should be the "order form" of the thing purchased and will tell you details about your order. The order should also be available in your account. Finally, there is a forms.py which will hold some forms models. I read that it is good in web development to keep those seperate from the ones in models.py I guess? Supposedly helps with optimization and doesn't chunk the code up as much in one file. That is all for now.